move db away from api directory
