+++ backend/api/types/schema.ts
@@ -1,1 +1,1 @@
-export * from '../db/schema'
+export * from '../../db/schema'
