+++ backend/db/schema/courseRatings.ts
