+++ backend/db/schema/courses.ts
