+++ backend/package.json
@@ -22,11 +22,11 @@
     "lint": "eslint .",
     "deploy": "npm run build && npm start",
     "dev": "nodemon",
-    "d:generate-schema": "npx drizzle-kit introspect --config api/db/config/drizzle.ts",
-    "d:migrate": "npx drizzle-kit migrate --config api/db/config/drizzle.ts",
-    "d:rollback": "npx drizzle-kit rollback --config api/db/config/drizzle.ts",
-    "d:studio": "npx drizzle-kit studio --config api/db/config/drizzle.ts",
-    "d:push": "npx drizzle-kit push --config api/db/config/drizzle.ts",
+    "d:generate-schema": "npx drizzle-kit introspect --config db/config/drizzle.ts",
+    "d:migrate": "npx drizzle-kit migrate --config db/config/drizzle.ts",
+    "d:rollback": "npx drizzle-kit rollback --config db/config/drizzle.ts",
+    "d:studio": "npx drizzle-kit studio --config db/config/drizzle.ts",
+    "d:push": "npx drizzle-kit push --config db/config/drizzle.ts",
     "vercel-build": "echo"
   },
   "devDependencies": {
